+++ app/components/Card.js
@@ -2,19 +2,19 @@ import React from 'react';
 import { View, StyleSheet, Image } from 'react-native';
 
 import colors from '../config/colors';
-import AppText from './AppText';
+import Text from './Text';
 
 function Card({ title, subTitle, image }) {
   return (
     <View style={styles.card}>
       <Image style={styles.image} source={image} />
       <View style={styles.detailsContainer}>
-        <AppText style={styles.title} numberOfLines={1}>
+        <Text style={styles.title} numberOfLines={1}>
           {title}
-        </AppText>
-        <AppText style={styles.subTitle} numberOfLines={2}>
+        </Text>
+        <Text style={styles.subTitle} numberOfLines={2}>
           {subTitle}
-        </AppText>
+        </Text>
       </View>
     </View>
   );
