+++ app/components/CategoryPickerItem.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
+
 import Icon from './Icon';
-import AppText from './AppText';
+import Text from './Text';
 
 function CategoryPickerItem({ item, onPress }) {
   return (
     <View style={styles.container}>
       <Icon backgroundColor={item.backgroundColor} name={item.icon} size={80} />
-      <AppText style={styles.label}>{item.label}</AppText>
+      <Text style={styles.label}>{item.label}</Text>
     </View>
   );
 }
