+++ app/components/Text.js
