+++ app/components/forms/FormField.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormikContext } from 'formik';
 
-import AppTextInput from '../AppTextInput';
+import TextInput from '../TextInput';
 import ErrorMessage from './ErrorMessage';
 
 function AppFormField({ name, width, ...otherProps }) {
@@ -9,7 +9,7 @@ function AppFormField({ name, width, ...otherProps }) {
 
   return (
     <>
-      <AppTextInput
+      <TextInput
         onBlur={() => setFieldTouched(name)}
         onChangeText={handleChange(name)}
         width={width}
