+++ app/components/forms/SubmitButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormikContext } from 'formik';
 
-import AppButton from '../AppButton';
+import AppButton from '../Button';
 
 function SubmitButton({ title }) {
   const { handleSubmit } = useFormikContext();
