+++ app/components/forms/index.js
@@ -1,5 +1,5 @@
-export { default as AppForm } from './AppForm';
-export { default as AppFormField } from './AppFormField';
+export { default as AppForm } from './Form';
+export { default as AppFormField } from './FormField';
 export { default as ErrorMessage } from './ErrorMessage';
 export { default as SubmitButton } from './SubmitButton';
-export { default as AppFormPicker } from './AppFormPicker';
+export { default as AppFormPicker } from './FormPicker';
