+++ app/screens/ListingDetailsScreen.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { View, Image, StyleSheet } from 'react-native';
 
-import AppText from '../components/AppText';
 import colors from '../config/colors';
 import { ListItem } from '../components/lists/';
+import Text from '../components/Text';
 
 function ListingDetailsScreen(props) {
   return (
     <View>
       <Image style={styles.image} source={require('../assets/jacket.jpg')} />
       <View style={styles.detailsContainer}>
-        <AppText style={styles.title}>Red jacket for sale</AppText>
-        <AppText style={styles.price}>$100</AppText>
+        <Text style={styles.title}>Red jacket for sale</Text>
+        <Text style={styles.price}>$100</Text>
         <View style={styles.userContainer}>
           <ListItem
             image={require('../assets/JasonPF.jpg')}
