+++ app/screens/WelcomeScreen.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ImageBackground, StyleSheet, View, Image, Text } from 'react-native';
-import AppButton from '../components/AppButton';
+import AppButton from '../components/Button';
 
 function WelcomeScreen(props) {
   return (
